restructuring parameters adding execute
