+++ assembly/compliance-tool/src/main/java/org/eclipse/sw360/antenna/frontend/compliancetool/sw360/status/SW360StatusReporterParameters.java
@@ -4,15 +4,13 @@ import org.eclipse.sw360.antenna.frontend.compliancetool.main.AntennaComplianceT
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.Collections;
-import java.util.HashSet;
 import java.util.Optional;
 import java.util.Set;
 
 public class SW360StatusReporterParameters {
     private static final Logger LOGGER = LoggerFactory.getLogger(SW360StatusReporterParameters.class);
 
-    private static final String REPORTER_PARAMETER_PREFIX = "--info=";
+    public static final String REPORTER_PARAMETER_PREFIX = "--info=";
 
     public InfoParameter getInfoRequestFromParameter(Set<String> parameters) {
         if (parameters.isEmpty()) {
@@ -30,18 +28,22 @@ public class SW360StatusReporterParameters {
 
     private InfoParameter parse(Set<String> parameters) {
         final Optional<String> infoParameter = parameters.stream().filter(p -> p.contains(REPORTER_PARAMETER_PREFIX)).findFirst();
-        final Optional<InfoParameter> optionalInfoParameter = infoParameter.map(this::getInfoParameterFromString);
+        final InfoParameter infoParameter1 = infoParameter.map(this::getInfoParameterFromString).orElse(emptyInfoParameter());
 
-        return optionalInfoParameter.orElse(null);
+        if (infoParameter1 == emptyInfoParameter()) {
+            throw new IllegalArgumentException(infoParameter.get() + ": " + infoParameter1.helpMessage());
+        }
+
+        return infoParameter1;
     }
 
-    private InfoParameter getInfoParameterFromString (String infoParameter) {
+    private InfoParameter getInfoParameterFromString(String infoParameter) {
         InfoParameter infoParam;
         switch (infoParameter) {
-            case new GetReleases().getInfoParameter():
-                return new GetReleases();
-            case new GetReleasesOfProjects().getInfoParameter():
-                return new GetReleasesOfProjects();
+            case IPGetReleases.getInfoParameter():
+                return new IPGetReleases();
+            case IPGetReleasesOfProjects.getInfoParameter():
+                return new IPGetReleasesOfProjects();
             default:
                 return emptyInfoParameter();
 
@@ -64,16 +66,7 @@ public class SW360StatusReporterParameters {
                 ":   Sets log level to DEBUG for diagnostic purposes." + cr + cr +
                 "The reporter create a csv file for every run with the requested information statement. " + cr + cr +
                 "The reporter info statements: (only one can be set)" + cr +
-                new GetReleases().getInfoParameter() + ":   Gives a list of all releases in a given sw360 instances that are cleared." + cr;
-    }
-
-    abstract class InfoParameter {
-        Set<String> parameters;
-        abstract String getInfoParameter();
-        abstract boolean hasAdditionalParameters();
-        abstract String helpMessage();
-        abstract boolean isValid();
-        abstract Set<String> getAdditionalParameters();
+                new IPGetReleases().getInfoParameter() + ":   Gives a list of all releases in a given sw360 instances that are cleared." + cr;
     }
 
     InfoParameter emptyInfoParameter() {
@@ -102,77 +95,11 @@ public class SW360StatusReporterParameters {
             Set<String> getAdditionalParameters() {
                 return null;
             }
+
+            @Override
+            void execute() {
+                // no-op
+            }
         };
-    }
-
-    class GetReleases extends InfoParameter{
-        static final String GET_RELEASES_CLEARED = REPORTER_PARAMETER_PREFIX + "releases-cleared";
-
-        @Override
-        static String getInfoParameter() {
-            return GET_RELEASES_CLEARED;
-        }
-
-        @Override
-        boolean hasAdditionalParameters() {
-            return false;
-        }
-
-        @Override
-        String helpMessage() {
-            return null;
-        }
-
-        @Override
-        boolean isValid() {
-            return false;
-        }
-
-        @Override
-        Set<String> getAdditionalParameters() {
-            return Collections.emptySet();
-        }
-    }
-
-    class GetReleasesOfProjects extends InfoParameter {
-        static final String GET_RELEASES_OF_PROJECT = REPORTER_PARAMETER_PREFIX + "releases-of-project";
-        static final String PROJECT_NAME = "--project_name";
-        static final String PROJECT_ID = "--project_id";
-
-        @Override
-        String getInfoParameter() {
-            return GET_RELEASES_OF_PROJECT;
-        }
-
-        @Override
-        boolean hasAdditionalParameters() {
-            return true;
-        }
-
-        @Override
-        String helpMessage() {
-            return null;
-        }
-
-        @Override
-        boolean isValid() {
-            return false;
-        }
-
-        @Override
-        Set<String> getAdditionalParameters() {
-            Set<String> additionalParamaters = new HashSet<>();
-            additionalParamaters.add(PROJECT_ID);
-            additionalParamaters.add(PROJECT_NAME);
-            return additionalParamaters;
-        }
-
-        public String getProjectId() {
-            return PROJECT_ID;
-        }
-
-        public String getProjectName() {
-            return PROJECT_NAME;
-        }
     }
 }
